feat: Add collections configuration for pages
This commit adds the necessary configuration for the "pages" collection in the "_config.yml" file. The "output" property is set to true, indicating that the pages in this collection should be outputted. The "sort_by" property is set to "date" to specify the sorting order of the pages. Additionally, the "permalink" property is defined to specify the URL structure for the pages in the collection. These changes were made to enable proper handling and rendering of the pages collection.